requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.1
numpy==1.26.0
requests==2.31.0
matplotlib==3.8.0
scikit-learn==1.3.1
